- Small refactor
